+++ docs/source/strategies.rst
@@ -14,31 +14,14 @@ The abstract interface to be implemented when creating a new Processing Strategy
    :undoc-members:
    :show-inheritance:
 
-Transformers
------------------------------
 
-Transformation steps. They transform the messages one by one provided a processing
-function.
-
-.. automodule:: arroyo.processing.strategies.transform
-   :members:
-   :undoc-members:
-
-Filters
+Filter
 -----------------------------
 
 .. automodule:: arroyo.processing.strategies.filter
    :members:
    :undoc-members:
 
-Batch and Unbatch
------------------------------
-
-Accumulate messages into a batch and pass to the next step.
-
-.. automodule:: arroyo.processing.strategies.batching
-   :members:
-   :undoc-members:
 
 Reduce (Fold)
 -----------------------------
@@ -58,12 +41,54 @@ Generates a sequence of messages from a single message based on a custom generat
    :members:
    :undoc-members:
 
-Task Runners
+
+Batch and Unbatch
+-----------------------------
+
+Accumulate messages into a batch and pass to the next step.
+The batch and unbatch strategies are based on reduce and unfold.
+Use reduce/unfold instead if you want to provide custom
+accumulator/generator functions.
+
+.. automodule:: arroyo.processing.strategies.batching
+   :members:
+   :undoc-members:
+
+
+Run Task
 -----------------------------
 
 .. automodule:: arroyo.processing.strategies.run_task
    :members:
    :undoc-members:
+
+
+Run Task in Threads
+-----------------------------
+
+.. automodule:: arroyo.processing.strategies.run_task_in_threads
+   :members:
+   :undoc-members:
+
+
+Run Task with Multiprocessing
+-----------------------------
+
+.. automodule:: arroyo.processing.strategies.run_task_with_multiprocessing
+   :members:
+   :undoc-members:
+
+
+Transformers
+-----------------------------
+
+Transformation steps. They transform the messages one by one provided a processing
+function. Alias for RunTask strategies.
+
+.. automodule:: arroyo.processing.strategies.transform
+   :members:
+   :undoc-members:
+
 
 Producers
 -----------------------------
@@ -71,6 +96,7 @@ Producers
 .. automodule:: arroyo.processing.strategies.produce
    :members:
    :undoc-members:
+
 
 Decoders
 -----------------------------
@@ -92,3 +118,13 @@ ground up in an upcoming release.
    :members:
    :undoc-members:
    :show-inheritance:
+
+Commit offsets
+-----------------------------
+
+Should be used as the last strategy in the chain, to ensure
+that offsets are only committed once all processing is complete.
+
+.. automodule:: arroyo.processing.strategies.commit
+   :members:
+   :undoc-members:
